requirements.txt
# Core dependencies (keeping existing)
numpy>=1.19.2
opencv-python>=4.5.1
matplotlib>=3.3.4
scikit-image>=0.17.2
scipy>=1.6.0
scikit-learn>=0.24.1
Pillow>=8.1.0

# FastAPI dependencies (new)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gunicorn>=21.2.0

# Remove streamlit and requests (not needed)